+++ config/config.exs
@@ -1,3 +1,0 @@
-use Mix.Config
-
-config :logger, :console, level: :info