+++ test/test_helper.exs
@@ -2,6 +2,7 @@
 Mix.start()
 Mix.shell(Mix.Shell.Process)
 System.put_env("ECTO_EDITOR", "")
+Logger.configure(level: :info)
 
 # Commonly used support feature
 Code.require_file "support/file_helpers.exs", __DIR__